# GGRC-1690 Add last_assessment_date to Control and Objective