--- a/test/unit/ggrc/converters/handlers/test_custom_attribute.py
+++ b/test/unit/ggrc/converters/handlers/test_custom_attribute.py
@@ -9,6 +9,7 @@ import unittest
 
 from mock import MagicMock, patch
 
+from ggrc import app  # noqa  # pylint: disable=unused-import
 from ggrc.converters.handlers.custom_attribute import (
     CustomAttributeColumHandler
 )