Note: performance of Control and Objective serialization suffers from this new property.

Objects | Time before | Time after
--------|-------------|-----------
10 Controls | 351 (std: 83.31) | 557 (std: 100.93)
20 Controls | 454 (std: 88.15) | 779 (std: 100.56)
100 Controls | 1159 (std: 75.00) | 2121 (std: 85.54)
500 Controls | 5021 (std: 266.46) | 9386 (std: 324.85)
2000 Controls | 11276 (std: 370.48) | 19035 (std: 573.96)
10 Objectives | 147 (std: 8.70) | 179 (std: 16.37)
20 Objectives | 180 (std: 10.79) | 205 (std: 21.60)
100 Objectives | 519 (std: 73.60) | 564 (std: 97.39)
500 Objectives | 2303 (std: 83.50) | 2366 (std: 126.34)
2000 Objectives | 3230 (std: 93.18) | 3335 (std: 120.22)

The table shows the average and std from 10 iterations.

~Note: Indexing logic uses util from #5421, that's why everything fails.~