@alieh-rymasheuski 2017-04-05
`GGRC-1780` - a ticket to re-implement this feature using a separate column.

By Nastassia: we can merge this PR now.