@NastassiaKiseleva 2017-04-06
@zidarsk8 @alieh-rymasheuski Just discussed this with Akhil - we have decided to push it for now - and work on a fix next sprint (please insert a ticket# here)